requirements.txt
absl-py==0.13.0
allennlp==2.4.0
arrow==1.1.1
astunparse==1.6.3
attrs==21.2.0
bert-score==0.3.10
blis==0.7.4
boto3==1.18.36
botocore==1.21.36
bpemb==0.3.3
cached-property==1.5.2
cachetools==4.2.2
catalogue==2.0.6
certifi==2021.5.30
cffi==1.14.5
charset-normalizer==2.0.4
clang==5.0
click==7.1.2
cloudpickle==1.6.0
configparser==5.0.2
conllu==4.4.1
cycler==0.10.0
cymem==2.0.5
datasets==1.11.0
Deprecated==1.2.13
dill==0.3.4
distro-info==0.21
docker-pycreds==0.4.0
docopt==0.6.2
editdistance==0.5.3
filelock==3.0.12
flair==0.9
flatbuffers==1.12
fsspec==2021.8.1
ftfy==6.0.3
future==0.18.2
gast==0.4.0
gdown==3.12.2
gensim==3.8.3
gitdb==4.0.7
GitPython==3.1.20
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.39.0
h5py==3.4.0
horovod==0.22.0
huggingface-hub==0.0.16
hyperopt==0.2.5
idna==3.2
importlib-metadata==3.10.1
iniconfig==1.1.1
iotop==0.6
Janome==0.4.1
jieba==0.42.1
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
jsonnet==0.17.0
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
konoha==4.6.5
langdetect==1.0.9
language-tool-python==2.5.5
lemminflect==0.2.2
lmdb==1.2.1
lru-dict==1.1.7
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
more-itertools==8.8.0
mpld3==0.3
multiprocess==0.70.12.2
murmurhash==1.0.5
networkx==2.6.2
nltk==3.6.2
num2words==0.5.10
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
overrides==3.1.0
packaging==21.0
pandas==1.3.2
pathtools==0.1.2
pathy==0.6.0
Pillow==8.2.0
pluggy==1.0.0
preshed==3.0.5
promise==2.3
protobuf==3.17.3
psutil==5.8.0
py==1.10.0
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycurl==7.43.0.2
pydantic==1.8.2
PyGObject==3.30.4
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.5
python-apt==1.8.4.3
python-dateutil==2.8.2
pytorch-crf==0.7.0
pytorch-nlp==0.5.0
pytz==2021.1
PyYAML==5.4.1
regex==2021.8.28
requests==2.26.0
requests-oauthlib==1.3.0
revtok==0.0.3
rsa==4.7.2
s3transfer==0.5.0
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.7.1
segtok==1.5.10
sentencepiece==0.1.95
sentry-sdk==1.3.1
shortuuid==1.0.1
six==1.16.0
smart-open==5.2.1
smmap==4.0.0
spacy==3.0.7
spacy-legacy==3.0.8
sqlitedict==1.7.0
srsly==2.4.1
subprocess32==3.5.4
tabulate==0.8.9
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4
tensorflow==2.6.0
tensorflow-estimator==2.6.0
tensorflow-hub==0.12.0
termcolor==1.1.0
terminaltables==3.1.0
thinc==8.0.9
threadpoolctl==2.2.0
tokenizers==0.10.3
toml==0.10.2
torch==1.8.0
torchaudio==0.8.1
torchtext==0.4.0
torchvision==0.9.1+cu111
tqdm==4.49.0
transformers==4.5.1
typer==0.3.2
typing-extensions==3.7.4.3
unattended-upgrades==0.1
urllib3==1.26.6
wandb==0.10.33
wasabi==0.8.2
wcwidth==0.2.5
Werkzeug==2.0.1
Wikipedia-API==0.5.4
word2number==1.1
wrapt==1.12.1
xxhash==2.0.2
zipp==3.5.0

